[Update] Move the initial state object to src/index.js
