+++ src/index.js
@@ -3,29 +3,40 @@ import ReactDOM from 'react-dom';
 import { Provider, connect } from 'react-redux';
 import { createStore } from 'redux';
 import App from './components/app';
-import { CREATE_BOOK, REMOVE_BOOK } from './actions/index';
+import { mapDispatchToProps } from './containers/booksList';
 import rootReducer from './reducers/index';
 
-const mapStateToProps = state => ({
+
+const randId = () => Math.floor(Math.random() * 1000);
+const defaultState = [
+  {
+    bookId: randId(),
+    title: 'First book',
+    category: 'Action',
+  },
+  {
+    bookId: randId(),
+    title: 'Learning React',
+    category: 'Learning',
+  },
+  {
+    bookId: randId(),
+    title: 'It chapter 2',
+    category: 'Horror',
+  },
+];
+
+const mapStateToProps = (state = defaultState) => ({
   books: state,
 });
 
-const mapDispatchToProps = dispatch => ({
-  addBook: book => {
-    dispatch(CREATE_BOOK(book));
-  },
-  removeBook: book => {
-    dispatch(REMOVE_BOOK(book));
-  },
-});
-
-
 const store = createStore(rootReducer);
-const Container = connect(mapStateToProps, mapDispatchToProps)(App);
+connect(mapStateToProps, mapDispatchToProps)(App);
+
 
 ReactDOM.render(
   <Provider store={store}>
-    <Container />
+    <App />
   </Provider>,
   document.getElementById('root'),
 );
